Added better UX and prepare a new function "Peter guess the number"
A function has been added to the code that displays the remaining range when searching for the next number. Second function has been prepared that allows the player to choose between 2 game options. The new function will allow the player to think of a number that the computer will try to guess within a few tries. In this version, the player has the option to choose this game mode, but will receive a message "coming soon."